LPS-167967 Move ObjectRelationshipUtil to package com.liferay.object.relationship.util
